requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
django-friendship==1.8.0
django-registration==2.4.1
django-tinymce==2.8.0
more-itertools==6.0.0
pep8==1.7.1
Pillow==5.4.1
pluggy==0.9.0
psycopg2==2.7.7
py==1.8.0
pycodestyle==2.5.0
pytest==4.3.0
python-decouple==3.1
pytz==2018.9
six==1.12.0
